==> 5. includeFunding.R <==
# Include funding formula
# Should have data here: https://www.cde.ca.gov/fg/fr/eb/
# Individual school files are here: http://lcapwatch.org/
# Yearly SARC files are like 2016: https://www.cde.ca.gov/ta/ac/sa/accessdata1617.asp
# Could create theoretical values from: https://www.cde.ca.gov/fg/aa/pa/ratesandinfo.asp, but they change each year!
# Check theoretical against averages: https://www.cde.ca.gov/fg/fr/sa/cefavgsalaries.asp

# Still Collecting data!
